fixed type of param value

+++ src/modules/vtol_att_control/colugoTransHelper.cpp
@@ -309,7 +309,7 @@ void colugoTransHelper::findAileronFuncs()
 void colugoTransHelper::setAsElevator(){
     char buffer[17];
         snprintf(buffer, sizeof(buffer), "CA_SV_CS%u_TYPE", _ailerons_tr_colugo._leftAileronCsTypeNo);
-        uint8_t val = static_cast<uint8_t>(ActuatorEffectivenessControlSurfaces::Type::Elevator);
+        int32_t val = static_cast<uint8_t>(ActuatorEffectivenessControlSurfaces::Type::Elevator);
         param_set(param_find(buffer), &val);
         snprintf(buffer, sizeof(buffer), "CA_SV_CS%u_TYPE", _ailerons_tr_colugo._rightAileronCsTypeNo);
       //  val = static_cast<uint8_t>(ActuatorEffectivenessControlSurfaces::Type::Elevator);
@@ -331,7 +331,7 @@ void colugoTransHelper::setAsElevator(){
 void colugoTransHelper::setAsAilerons(){
     char buffer[17];
         snprintf(buffer, sizeof(buffer), "CA_SV_CS%u_TYPE", _ailerons_tr_colugo._leftAileronCsTypeNo);
-        uint8_t val = static_cast<uint8_t>(ActuatorEffectivenessControlSurfaces::Type::LeftAileron);
+        int32_t val = static_cast<uint8_t>(ActuatorEffectivenessControlSurfaces::Type::LeftAileron);
         param_set(param_find(buffer), &val);
         snprintf(buffer, sizeof(buffer), "CA_SV_CS%u_TYPE", _ailerons_tr_colugo._rightAileronCsTypeNo);
         val = static_cast<uint8_t>(ActuatorEffectivenessControlSurfaces::Type::RightAileron);
